refacotr: seperate customGrid from grid
